Inital experiments
While running your script Rafi, I found a pretty nasty bug in ao_core that meant your results with this experiment were invalid-- I'm glad we caught it before it hurt anyone.

Then I ran your experiment and it is now performing well. Aside from fixing the ao_core related bug, I added reset states and set in unsequenced=True in next_state, as is logical with this use-case. Though it'd be SUPER interesting to have an ao agent that outputs embeddings/tokens, with the default of sequence, to see if it can "write" sentences. Let's continue these experiments!

Co-Authored-By: Rafayel Latif <74052861+Rafipilot@users.noreply.github.com>